+++ qualtran/bloqs/chemistry/thc/prepare.py
@@ -45,7 +45,7 @@ from qualtran.bloqs.basic_gates.on_each import OnEach
 from qualtran.bloqs.block_encoding.lcu_select_and_prepare import PrepareOracle
 from qualtran.bloqs.data_loading.select_swap_qrom import SelectSwapQROM
 from qualtran.bloqs.mcmt.multi_control_multi_target_pauli import MultiControlPauli
-from qualtran.bloqs.reflection import Reflection
+from qualtran.bloqs.reflections import Reflection
 from qualtran.cirq_interop import CirqGateAsBloq
 from qualtran.drawing import Text, WireSymbol
 from qualtran.linalg.lcu_util import preprocess_lcu_coefficients_for_reversible_sampling
